Update analytics.py
following up on the instructions last time, I added W as an optional parameter for the method. If W is not passed, then libpysal.weights.distance.Kernel is applied to compute the weight. fill_diagonal function is applied to fill the diagonal. I also added row_standardize as an optional parameter, determining whether or not to standardize the rows in the weight matrix